+++ CHANGELOG.md
@@ -2,8 +2,10 @@
 
 ## 0.0.5
 * Add Function method `debounce` `throttle` `curry`
-* Add `hex2Rgba` method
+* Add `hex2Rgba` `validate` method
 * Add `limitNumber` `rand` `toThousands` method
+* Add `objects` module include `omit` `pick` `filterEmpty` method
+* Add `strings` module include `randomString` `upFirstLetter` `lowFirstLetter` method
 
 ## 0.0.4
 * Add `sliceLoad` method